feat: Improvements on DaemonSet controller
