+++ control-plane-operator/hostedclusterconfigoperator/controllers/globalps/globalps.go
@@ -2,35 +2,35 @@ package globalps
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"os"
-	"strings"
-	"time"
 
-	resourceglobalps "github.com/openshift/hypershift/control-plane-operator/hostedclusterconfigoperator/controllers/resources/globalps"
 	"github.com/openshift/hypershift/control-plane-operator/hostedclusterconfigoperator/controllers/resources/manifests"
+	"github.com/openshift/hypershift/support/thirdparty/kubernetes/pkg/credentialprovider"
 	"github.com/openshift/hypershift/support/upsert"
 
+	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
+	rbacv1 "k8s.io/api/rbac/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	kubeclient "k8s.io/client-go/kubernetes"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/utils/ptr"
 
 	ctrl "sigs.k8s.io/controller-runtime"
 	crclient "sigs.k8s.io/controller-runtime/pkg/client"
 	crreconcile "sigs.k8s.io/controller-runtime/pkg/reconcile"
+
+	"github.com/go-logr/logr"
 )
 
 const (
 	ControllerName = "globalps"
 )
 
-var (
-	recoverBeforeShutdown = true
-)
-
 type Reconciler struct {
 	cpClient               crclient.Client
-	hcKubeClient           kubeclient.Interface
 	kubeSystemSecretClient crclient.Client
 	hcUncachedClient       crclient.Client
 	hcpNamespace           string
@@ -48,10 +48,7 @@ func (r *Reconciler) Reconcile(ctx context.Context, req crreconcile.Request) (cr
 	}
 
 	if err := r.reconcileGlobalPullSecret(ctx, r.hcpNamespace, hccoImage); err != nil {
-		if !strings.Contains(err.Error(), "global pull secret syncer signaled to shutdown") {
-			return ctrl.Result{}, err
-		}
-		return ctrl.Result{RequeueAfter: 30 * time.Second}, nil
+		return ctrl.Result{}, fmt.Errorf("failed to reconcile global pull secret: %w", err)
 	}
 
 	return ctrl.Result{}, nil
@@ -69,62 +66,26 @@ func (r *Reconciler) reconcileGlobalPullSecret(ctx context.Context, hcpNamespace
 		err                         error
 	)
 	log := ctrl.LoggerFrom(ctx)
-	log.Info("Reconciling global pull secret")
+	log.Info("reconciling global pull secret")
 
 	// Get the user provided pull secret
-	exists, additionalPullSecret, err := resourceglobalps.AdditionalPullSecretExists(ctx, r.kubeSystemSecretClient)
+	exists, additionalPullSecret, err := AdditionalPullSecretExists(ctx, r.kubeSystemSecretClient)
 	if err != nil {
 		return fmt.Errorf("failed to check if user provided pull secret exists: %w", err)
 	}
 
 	// Reconcile the RBAC for the Global Pull Secret
-	if err := resourceglobalps.ReconcileGlobalPullSecretRBAC(ctx, r.hcUncachedClient, r.kubeSystemSecretClient, r.CreateOrUpdate, hcpNamespace); err != nil {
+	if err := ReconcileGlobalPullSecretRBAC(ctx, r.hcUncachedClient, r.kubeSystemSecretClient, r.CreateOrUpdate, hcpNamespace); err != nil {
 		return fmt.Errorf("failed to reconcile global pull secret RBAC: %w", err)
 	}
 
 	if !exists {
-		// Early cleanup
-		if recoverBeforeShutdown {
-			// Recover the original pull secret
-			log.Info("Recovering original pull secret")
-			originalPullSecret := manifests.PullSecret(hcpNamespace)
-			if err := r.cpClient.Get(ctx, crclient.ObjectKeyFromObject(originalPullSecret), originalPullSecret); err != nil {
-				return fmt.Errorf("failed to get original pull secret: %w", err)
-			}
-			originalPullSecretBytes = originalPullSecret.Data[corev1.DockerConfigJsonKey]
-
-			// Create secret in the DataPlane
-			secret := manifests.GlobalPullSecret()
-			if _, err := r.CreateOrUpdate(ctx, r.kubeSystemSecretClient, secret, func() error {
-				secret.Data = map[string][]byte{
-					corev1.DockerConfigJsonKey: originalPullSecretBytes,
-				}
-				return nil
-			}); err != nil {
-				return fmt.Errorf("failed to create global pull secret: %w", err)
-			}
-			log.Info("Original pull secret recovered, global pull secret syncer signaled to shutdown")
-			recoverBeforeShutdown = false
-
-			return fmt.Errorf("original pull secret recovered, global pull secret syncer signaled to shutdown")
-		}
-
-		// Delete the global pull secret and the daemon set
 		secret := manifests.GlobalPullSecret()
 		if err := r.kubeSystemSecretClient.Delete(ctx, secret); err != nil {
 			if !apierrors.IsNotFound(err) {
 				return fmt.Errorf("failed to delete global pull secret: %w", err)
 			}
 		}
-
-		daemonSet := manifests.GlobalPullSecretDaemonSet()
-		if err := r.hcUncachedClient.Delete(ctx, daemonSet); err != nil {
-			if !apierrors.IsNotFound(err) {
-				return fmt.Errorf("failed to delete global pull secret daemon set: %w", err)
-			}
-		}
-
-		log.Info("Skipping global pull secret reconciliation")
 		return nil
 	}
 
@@ -133,12 +94,11 @@ func (r *Reconciler) reconcileGlobalPullSecret(ctx context.Context, hcpNamespace
 		return nil
 	}
 
-	if userProvidedPullSecretBytes, err = resourceglobalps.ValidateAdditionalPullSecret(additionalPullSecret); err != nil {
+	if userProvidedPullSecretBytes, err = ValidateAdditionalPullSecret(additionalPullSecret); err != nil {
 		return fmt.Errorf("failed to validate user provided pull secret: %w", err)
 	}
 
 	log.Info("Valid additional pull secret found in the DataPlane, reconciling global pull secret")
-	recoverBeforeShutdown = true
 
 	// Get the original pull secret
 	originalPullSecret := manifests.PullSecret(hcpNamespace)
@@ -150,7 +110,7 @@ func (r *Reconciler) reconcileGlobalPullSecret(ctx context.Context, hcpNamespace
 	originalPullSecretBytes = originalPullSecret.Data[corev1.DockerConfigJsonKey]
 
 	// Merge the additional pull secret with the original pull secret
-	if globalPullSecretBytes, err = resourceglobalps.MergePullSecrets(ctx, originalPullSecretBytes, userProvidedPullSecretBytes); err != nil {
+	if globalPullSecretBytes, err = MergePullSecrets(ctx, originalPullSecretBytes, userProvidedPullSecretBytes); err != nil {
 		return fmt.Errorf("failed to merge pull secrets: %w", err)
 	}
 
@@ -167,9 +127,240 @@ func (r *Reconciler) reconcileGlobalPullSecret(ctx context.Context, hcpNamespace
 
 	// Use the Global Pull Secret to deploy the DaemonSet in the DataPlane.
 	daemonSet := manifests.GlobalPullSecretDaemonSet()
-	if err := resourceglobalps.ReconcileDaemonSet(ctx, daemonSet, secret.Name, r.hcUncachedClient, r.CreateOrUpdate, cpoImage); err != nil {
+	if err := ReconcileDaemonSet(ctx, daemonSet, secret.Name, r.hcUncachedClient, r.CreateOrUpdate, cpoImage); err != nil {
 		return fmt.Errorf("failed to reconcile global pull secret daemon set: %w", err)
 	}
 
 	return nil
+}
+
+func ReconcileDaemonSet(ctx context.Context, daemonSet *appsv1.DaemonSet, globalPullSecretName string, c crclient.Client, createOrUpdate upsert.CreateOrUpdateFN, cpoImage string) error {
+	log, err := logr.FromContext(ctx)
+	if err != nil {
+		return fmt.Errorf("failed to get logger: %w", err)
+	}
+
+	log.Info("Reconciling global pull secret daemon set")
+
+	if _, err := createOrUpdate(ctx, c, daemonSet, func() error {
+		daemonSet.Spec = appsv1.DaemonSetSpec{
+			Selector: &metav1.LabelSelector{
+				MatchLabels: map[string]string{
+					"name": manifests.GlobalPullSecretDSName,
+				},
+			},
+			Template: corev1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Labels: map[string]string{
+						"name": manifests.GlobalPullSecretDSName,
+					},
+				},
+				Spec: corev1.PodSpec{
+					ServiceAccountName:           manifests.GlobalPullSecretDSName,
+					AutomountServiceAccountToken: ptr.To(true),
+					SecurityContext:              &corev1.PodSecurityContext{},
+					DNSPolicy:                    corev1.DNSDefault,
+					Tolerations:                  []corev1.Toleration{{Operator: corev1.TolerationOpExists}},
+					Containers: []corev1.Container{
+						{
+							Name:            manifests.GlobalPullSecretDSName,
+							Image:           cpoImage,
+							ImagePullPolicy: corev1.PullIfNotPresent,
+							Command: []string{
+								"/usr/bin/control-plane-operator",
+							},
+							Args: []string{
+								"sync-global-pullsecret",
+								fmt.Sprintf("--global-pull-secret-name=%s", globalPullSecretName),
+							},
+							SecurityContext: &corev1.SecurityContext{
+								Privileged: ptr.To(true),
+							},
+							VolumeMounts: []corev1.VolumeMount{
+								{
+									Name:      "kubelet-config",
+									MountPath: "/var/lib/kubelet",
+								},
+								{
+									Name:      "dbus",
+									MountPath: "/var/run/dbus",
+								},
+							},
+							TerminationMessagePolicy: corev1.TerminationMessageFallbackToLogsOnError,
+							Resources: corev1.ResourceRequirements{
+								Requests: corev1.ResourceList{
+									corev1.ResourceMemory: resource.MustParse("50Mi"),
+									corev1.ResourceCPU:    resource.MustParse("40m"),
+								},
+							},
+						},
+					},
+					Volumes: []corev1.Volume{
+						{
+							Name: "kubelet-config",
+							VolumeSource: corev1.VolumeSource{
+								HostPath: &corev1.HostPathVolumeSource{
+									Path: "/var/lib/kubelet",
+									Type: ptr.To(corev1.HostPathDirectory),
+								},
+							},
+						},
+						{
+							Name: "dbus",
+							VolumeSource: corev1.VolumeSource{
+								HostPath: &corev1.HostPathVolumeSource{
+									Path: "/var/run/dbus",
+									Type: ptr.To(corev1.HostPathDirectory),
+								},
+							},
+						},
+					},
+				},
+			},
+		}
+
+		return nil
+	}); err != nil {
+		return fmt.Errorf("failed to create global pull secret daemon set: %w", err)
+	}
+
+	return nil
+}
+
+func ValidateAdditionalPullSecret(pullSecret *corev1.Secret) ([]byte, error) {
+	var dockerConfigJSON credentialprovider.DockerConfigJSON
+
+	// Validate that the pull secret contains the dockerConfigJson key
+	if _, ok := pullSecret.Data[corev1.DockerConfigJsonKey]; !ok {
+		return nil, fmt.Errorf("pull secret data is not a valid docker config json")
+	}
+
+	// Validate that the pull secret is a valid DockerConfigJSON
+	pullSecretBytes := pullSecret.Data[corev1.DockerConfigJsonKey]
+	if err := json.Unmarshal(pullSecretBytes, &dockerConfigJSON); err != nil {
+		return nil, fmt.Errorf("invalid docker config json format: %w", err)
+	}
+
+	// Validate that the pull secret contains at least one auth entry
+	if len(dockerConfigJSON.Auths) == 0 {
+		return nil, fmt.Errorf("docker config json must contain at least one auth entry")
+	}
+
+	return pullSecretBytes, nil
+}
+
+// MergePullSecrets merges two pull secrets into a single pull secret.
+// The additional pull secret is merged with the original pull secret.
+// If an auth entry already exists, it will be overwritten.
+// The resulting pull secret is returned as a JSON string.
+// Not using credentialprovider.DockerConfigJSON because it does not support
+// marshaling the auth field.
+func MergePullSecrets(ctx context.Context, originalPullSecret, userProvidedPullSecret []byte) ([]byte, error) {
+	var (
+		originalAuths         map[string]any
+		userProvidedAuths     map[string]any
+		originalJSON          map[string]any
+		userProvidedJSON      map[string]any
+		globalPullSecretBytes []byte
+		err                   error
+	)
+
+	// Unmarshal original pull secret
+	if err = json.Unmarshal(originalPullSecret, &originalJSON); err != nil {
+		return nil, fmt.Errorf("invalid original pull secret format: %w", err)
+	}
+	originalAuths = originalJSON["auths"].(map[string]any)
+
+	// Unmarshal additional pull secret
+	if err = json.Unmarshal(userProvidedPullSecret, &userProvidedJSON); err != nil {
+		return nil, fmt.Errorf("invalid user provided pull secret format: %w", err)
+	}
+	userProvidedAuths = userProvidedJSON["auths"].(map[string]any)
+
+	// Merge auths
+	for k, v := range userProvidedAuths {
+		originalAuths[k] = v
+	}
+
+	// Create final JSON
+	finalJSON := map[string]any{
+		"auths": originalAuths,
+	}
+
+	globalPullSecretBytes, err = json.Marshal(finalJSON)
+	if err != nil {
+		return nil, fmt.Errorf("failed to marshal merged pull secret: %w", err)
+	}
+
+	return globalPullSecretBytes, nil
+}
+
+func ReconcileGlobalPullSecretRBAC(ctx context.Context, c crclient.Client, cacheClient crclient.Client, createOrUpdate upsert.CreateOrUpdateFN, namespace string) error {
+	// Remove the RBAC resources if the user provided pull secret is not present
+	log, err := logr.FromContext(ctx)
+	if err != nil {
+		return fmt.Errorf("failed to get logger: %w", err)
+	}
+
+	log.Info("reconciling global pull secret RBAC")
+
+	// Create ServiceAccount
+	sa := manifests.GlobalPullSecretServiceAccount()
+	if _, err := createOrUpdate(ctx, c, sa, func() error { return nil }); err != nil {
+		return fmt.Errorf("failed to reconcile service account: %w", err)
+	}
+
+	// Create cluster role with separate rules for list/watch and get
+	globalPullSecretRole := manifests.GlobalPullSecretRole()
+	if _, err := createOrUpdate(ctx, c, globalPullSecretRole, func() error {
+		globalPullSecretRole.Rules = []rbacv1.PolicyRule{
+			{
+				APIGroups: []string{""},
+				Resources: []string{"secrets"},
+				Verbs:     []string{"list", "watch"},
+			},
+			{
+				APIGroups:     []string{""},
+				Resources:     []string{"secrets"},
+				ResourceNames: []string{"additional-pull-secret", "global-pull-secret", "pull-secret"},
+				Verbs:         []string{"get"},
+			},
+		}
+		return nil
+	}); err != nil {
+		return fmt.Errorf("failed to reconcile global pull secret syncer role: %w", err)
+	}
+
+	// Create ClusterRoleBinding
+	globalPullSecretClusterRoleBinding := manifests.GlobalPullSecretRoleBinding()
+	if _, err := createOrUpdate(ctx, c, globalPullSecretClusterRoleBinding, func() error {
+		globalPullSecretClusterRoleBinding.RoleRef = rbacv1.RoleRef{
+			APIGroup: "rbac.authorization.k8s.io",
+			Kind:     "ClusterRole",
+			Name:     globalPullSecretRole.Name,
+		}
+		globalPullSecretClusterRoleBinding.Subjects = []rbacv1.Subject{
+			{
+				Kind:      "ServiceAccount",
+				Name:      sa.Name,
+				Namespace: sa.Namespace,
+			},
+		}
+		return nil
+	}); err != nil {
+		return fmt.Errorf("failed to reconcile global pull secret syncer role binding: %w", err)
+	}
+
+	return nil
+}
+
+func AdditionalPullSecretExists(ctx context.Context, c crclient.Client) (bool, *corev1.Secret, error) {
+	additionalPullSecret := manifests.AdditionalPullSecret()
+	if err := c.Get(ctx, crclient.ObjectKeyFromObject(additionalPullSecret), additionalPullSecret); err != nil {
+		if apierrors.IsNotFound(err) {
+			return false, nil, nil
+		}
+		return false, nil, err
+	}
+	return true, additionalPullSecret, nil
 }
