+++ control-plane-operator/hostedclusterconfigoperator/controllers/globalps/globalps_test.go
