+++ sync-global-pullsecret/sync-global-pullsecret.go
@@ -11,11 +11,14 @@ import (
 	hyperapi "github.com/openshift/hypershift/support/api"
 
 	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/rest"
 
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/cache"
-	"sigs.k8s.io/controller-runtime/pkg/client"
+	crclient "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
@@ -34,7 +37,10 @@ type syncGlobalPullSecretOptions struct {
 const (
 	defaultKubeletConfigJsonPath     = "/var/lib/kubelet/config.json"
 	defaultGlobalPSSecretName        = "global-pull-secret"
+	additionalPullSecretName         = "additional-pull-secret"
 	defaultGlobalPullSecretNamespace = "kube-system"
+	originalPullSecretName           = "pull-secret"
+	originalPullSecretNamespace      = "openshift-config"
 	dbusRestartUnitMode              = "replace"
 	kubeletServiceUnit               = "kubelet.service"
 
@@ -54,7 +60,8 @@ var writeFileFunc = os.WriteFile
 
 // GlobalPullSecretReconciler reconciles a Secret object
 type GlobalPullSecretReconciler struct {
-	client.Client
+	cachedClient          crclient.Client
+	uncachedClient        crclient.Client
 	Scheme                *runtime.Scheme
 	kubeletConfigJsonPath string
 	globalPSSecretName    string
@@ -102,9 +109,20 @@ func (o *syncGlobalPullSecretOptions) run(ctx context.Context) error {
 		return fmt.Errorf("failed to create manager: %w", err)
 	}
 
+	uncachedClientRestConfig := mgr.GetConfig()
+	uncachedClientRestConfig.WarningHandler = rest.NoWarnings{}
+	uncachedClient, err := crclient.New(uncachedClientRestConfig, crclient.Options{
+		Scheme: mgr.GetScheme(),
+		Mapper: mgr.GetRESTMapper(),
+	})
+	if err != nil {
+		return fmt.Errorf("failed to create uncached client: %w", err)
+	}
+
 	// Create reconciler
 	r := &GlobalPullSecretReconciler{
-		Client:                mgr.GetClient(),
+		cachedClient:          mgr.GetClient(),
+		uncachedClient:        uncachedClient,
 		Scheme:                mgr.GetScheme(),
 		kubeletConfigJsonPath: o.kubeletConfigJsonPath,
 		globalPSSecretName:    o.globalPSSecretName,
@@ -123,7 +141,7 @@ func (o *syncGlobalPullSecretOptions) run(ctx context.Context) error {
 				return o.isTargetSecret(e.ObjectNew)
 			},
 			DeleteFunc: func(e event.DeleteEvent) bool {
-				return false
+				return o.isTargetSecret(e.Object)
 			},
 			GenericFunc: func(e event.GenericEvent) bool {
 				return false
@@ -142,7 +160,7 @@ func (o *syncGlobalPullSecretOptions) run(ctx context.Context) error {
 }
 
 // isTargetSecret checks if the given object is the target secret we want to watch
-func (o *syncGlobalPullSecretOptions) isTargetSecret(obj client.Object) bool {
+func (o *syncGlobalPullSecretOptions) isTargetSecret(obj crclient.Object) bool {
 	// Check if it's a Secret and has the correct name and namespace
 	if secret, ok := obj.(*corev1.Secret); ok {
 		return secret.GetNamespace() == defaultGlobalPullSecretNamespace &&
@@ -153,28 +171,65 @@ func (o *syncGlobalPullSecretOptions) isTargetSecret(obj client.Object) bool {
 
 // Reconcile handles the reconciliation logic for the GlobalPullSecret
 func (r *GlobalPullSecretReconciler) Reconcile(ctx context.Context, req reconcile.Request) (reconcile.Result, error) {
-	log := ctrl.LoggerFrom(ctx)
-	log.Info("Reconciling GlobalPullSecret")
+	var (
+		log                = ctrl.LoggerFrom(ctx)
+		globalPullSecret   = &corev1.Secret{}
+		originalPullSecret = &corev1.Secret{}
+		err                error
+		chosenPullSecret   *corev1.Secret
+	)
 
-	// Get the global pull secret
-	globalPullSecret := &corev1.Secret{}
-	if err := r.Get(ctx, req.NamespacedName, globalPullSecret); err != nil {
-		log.Error(err, "Failed to get global pull secret")
-		return reconcile.Result{}, client.IgnoreNotFound(err)
+	log.Info("Reconciling GlobalPullSecret")
+	err = r.uncachedClient.Get(ctx, types.NamespacedName{
+		Name:      originalPullSecretName,
+		Namespace: originalPullSecretNamespace,
+	}, originalPullSecret)
+	if err != nil {
+		return reconcile.Result{}, fmt.Errorf("failed to get original pull secret: %w", err)
 	}
 
-	if err := r.checkAndFixFile(ctx, globalPullSecret); err != nil {
-		log.Error(err, "Failed to check and fix file")
-		return reconcile.Result{}, err
+	err = r.cachedClient.Get(ctx, req.NamespacedName, globalPullSecret)
+	if err != nil {
+		if !apierrors.IsNotFound(err) {
+			return reconcile.Result{}, fmt.Errorf("failed to get global pull secret: %w", err)
+		}
+		log.Info("Global pull secret not found, using original pull secret")
+		chosenPullSecret = originalPullSecret.DeepCopy()
+	} else {
+		log.Info("Global pull secret found, using it")
+		chosenPullSecret = globalPullSecret.DeepCopy()
+	}
+
+	// Normal reconciliation
+	if err := r.checkAndFixFile(ctx, chosenPullSecret); err != nil {
+		return reconcile.Result{}, fmt.Errorf("failed to check and fix file: %w", err)
 	}
 
 	return reconcile.Result{}, nil
 }
 
-// checkAndFixFile reads the current file content and updates it if it differs from the desired content.
+// checkAndFixFile reads the current file content and updates it if it differs from the desired content (global pull secret content).
 // Have in mind the logic which do the merge of the pull secret is in the globalpullsecret package under the HCCO.
-func (r *GlobalPullSecretReconciler) checkAndFixFile(ctx context.Context, globalPullSecret *corev1.Secret) error {
+func (r *GlobalPullSecretReconciler) checkAndFixFile(ctx context.Context, pullSecret *corev1.Secret) error {
 	log := ctrl.LoggerFrom(ctx)
+	log.Info("Checking and fixing file")
+
+	// Validate pullSecret is not nil
+	if pullSecret == nil {
+		return fmt.Errorf("pullSecret cannot be nil")
+	}
+
+	// Validate pullSecret.Data is not nil
+	if pullSecret.Data == nil {
+		return fmt.Errorf("pullSecret.Data cannot be nil")
+	}
+
+	log.Info("DEBUG: Pass Data check")
+	// Validate the required key exists
+	pullSecretBytes, exists := pullSecret.Data[corev1.DockerConfigJsonKey]
+	if !exists {
+		return fmt.Errorf("pullSecret does not contain required key: %s", corev1.DockerConfigJsonKey)
+	}
 
 	// Read existing content if file exists
 	existingContent, err := os.ReadFile(r.kubeletConfigJsonPath)
@@ -182,16 +237,14 @@ func (r *GlobalPullSecretReconciler) checkAndFixFile(ctx context.Context, global
 		return fmt.Errorf("failed to read existing file: %w", err)
 	}
 
-	// Get bytes of the global pull secret
-	globalPullSecretBytes := globalPullSecret.Data[corev1.DockerConfigJsonKey]
-
 	// If file content is different, write the desired content
-	if string(existingContent) != string(globalPullSecretBytes) {
+	if string(existingContent) != string(pullSecretBytes) {
+		log.Info("file content is different, updating it")
 		// Save original content for potential rollback
 		originalContent := existingContent
 
 		// Write the new content
-		if err := writeFileFunc(r.kubeletConfigJsonPath, globalPullSecretBytes, 0600); err != nil {
+		if err := writeFileFunc(r.kubeletConfigJsonPath, pullSecretBytes, 0600); err != nil {
 			return fmt.Errorf("failed to write file: %w", err)
 		}
 		log.Info("Pull secret updated", "file", r.kubeletConfigJsonPath)
